ul_map_backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
fancycompleter==0.9.1
idna==2.10
pdbpp==0.10.2
Pygments==2.8.1
pyreadline==2.1
pyrepl==0.9.0
python-dotenv==0.16.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
wmctrl==0.3
